Added dummy_data.py to the test directory and updated it to functions (instead of a script)
